# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: QAssess API
  version: 0.3.0
  description: |
    Multi-tenant assessment platform API supporting authoring, public runtime, scoring,
    reporting, analytics, and integrations.
servers:
  - url: https://api.qassess.example.com/v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /analytics/assessments/{assessmentId}/summary:
    get:
      tags: [Analytics]
      summary: Assessment funnel and score summary
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
  /analytics/assessments/{assessmentId}/dropoff:
    get:
      tags: [Analytics]
      summary: Question-level dropoff metrics
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Dropoff metrics
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DropoffMetric'
  /integrations/webhooks:
    get:
      tags: [Integrations]
      summary: List webhook endpoints
      responses:
        '200':
          description: Webhook endpoints
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEndpoint'
    post:
      tags: [Integrations]
      summary: Create webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointCreateRequest'
      responses:
        '201':
          description: Webhook endpoint created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
  /integrations/webhooks/{endpointId}:
    patch:
      tags: [Integrations]
      summary: Update webhook endpoint
      parameters:
        - $ref: '#/components/parameters/EndpointId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEndpointUpdateRequest'
      responses:
        '200':
          description: Webhook endpoint updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEndpoint'
    delete:
      tags: [Integrations]
      summary: Delete webhook endpoint
      parameters:
        - $ref: '#/components/parameters/EndpointId'
      responses:
        '204':
          description: Webhook endpoint deleted
  /assessments/{assessmentId}/leads/export:
    get:
      tags: [Integrations]
      summary: Export leads as CSV
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: CSV stream
          content:
            text/csv:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AssessmentId:
      in: path
      name: assessmentId
      required: true
      schema:
        type: string
        format: uuid
...
    EndpointId:
      in: path
      name: endpointId
      required: true
      schema:
        type: string
        format: uuid
    DateFrom:
      in: query
      name: dateFrom
      schema:
        type: string
        format: date
    DateTo:
      in: query
      name: dateTo
      schema:
        type: string
        format: date
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
    UserRole:
      type: string
      enum: [owner, editor, analyst, viewer]
...
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            fieldErrors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                  message:
                    type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        tenantSlug:
          type: string
          description: Optional when email exists in multiple tenants.
    AuthTokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, tenantId, email, role, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    AnalyticsSummary:
      type: object
      required: [assessmentId, dateFrom, dateTo, visits, starts, completions, leads, conversionRate]
      properties:
        assessmentId:
          type: string
          format: uuid
        dateFrom:
          type: string
          format: date
        dateTo:
          type: string
          format: date
        visits:
          type: integer
          minimum: 0
        starts:
          type: integer
          minimum: 0
        completions:
          type: integer
          minimum: 0
        leads:
          type: integer
          minimum: 0
        conversionRate:
          type: number
          minimum: 0
          maximum: 1
        averageScore:
          type: number
    DropoffMetric:
      type: object
      required: [questionId, views, exits, dropoffRate]
      properties:
        questionId:
          type: string
          format: uuid
        questionPrompt:
          type: string
        views:
          type: integer
          minimum: 0
        exits:
          type: integer
          minimum: 0
        dropoffRate:
          type: number
          minimum: 0
          maximum: 1
    WebhookEndpoint:
      type: object
      required: [id, tenantId, name, targetUrl, subscribedEvents, isActive]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        name:
          type: string
        targetUrl:
          type: string
          format: uri
        subscribedEvents:
          type: array
          items:
            type: string
            enum: [lead.created, session.completed, pdf.generated]
        isActive:
          type: boolean
    WebhookEndpointCreateRequest:
      type: object
      required: [name, targetUrl, secret, subscribedEvents]
      properties:
        name:
          type: string
        targetUrl:
          type: string
          format: uri
        secret:
          type: string
          minLength: 12
        subscribedEvents:
          type: array
          minItems: 1
          items:
            type: string
            enum: [lead.created, session.completed, pdf.generated]
    WebhookEndpointUpdateRequest:
      type: object
      properties:
        name:
          type: string
        targetUrl:
          type: string
          format: uri
        secret:
          type: string
          minLength: 12
        subscribedEvents:
          type: array
          items:
            type: string
            enum: [lead.created, session.completed, pdf.generated]
        isActive:
          type: boolean
...